chore: move arquivos para pages e components
