+++ frontend/src/components/sidebar/Sidebar.css
@@ -1,18 +1,64 @@
-#side{
-    width: 250px;
-    height: 100%;
-    background-color: #166fc938;
-    text-align: center;
-    
-
+body.sidebar-open {
+  overflow: hidden; /* Remove o scroll */
+  position: fixed; /* Fixa o fundo */
+  width: 100%; /* Garante largura fixa */
+  height: 100%; /* Garante altura fixa */
 }
 
-#side>nav>a{
-    display: block;
-    color: black;
-    font-size: 20px;
-    background-color: #166fc978;
-    margin-block: 1px;
-    padding: 10px;
-    
+.bm-burger-button {
+  position: absolute;
+  top: 30px;
+  right: 60px;
+  width: 36px;
+  height: 20px;
+  transition: 0.5s;
+}
+
+/* Color/shape of burger icon bars */
+.bm-burger-bars {
+  background: #373a47;
+}
+
+/* Position and sizing of clickable cross button */
+.bm-cross-button {
+  left: 20px;
+  height: 24px;
+  width: 24px;
+}
+
+/* Color/shape of close button cross */
+.bm-cross {
+  background: #373a47;
+}
+
+/* General sidebar styles */
+.bm-menu {
+  background: #c5dbf1;
+  padding: 3em 0;
+  margin-left: 0em;
+  font-size: 1.15em;
+  right: 0;
+}
+
+/* Wrapper for item list  */
+.bm-item-list {
+  text-align: center;
+  color: #b8b7ad;
+}
+
+.bm-item.menu-item {
+  width: 100%;
+  background-color: #99c0e7;
+  padding: 2em 0 2em 0;
+  margin: 0.2em 0 0.2em 0;
+  text-decoration: none;
+}
+
+.bm-item.menu-item:hover {
+  background-color: #b7cee4;
+}
+
+/* Styling of overlay  */
+.bm-overlay {
+  background: rgba(0, 0, 0, 0.3);
 }
