+++ frontend/src/components/sidebar/Sidebar.jsx
@@ -1,15 +1,37 @@
-import "./Sidebar.css"
-import { Link } from "react-router-dom"
+import React, { useEffect } from "react";
+import "./Sidebar.css";
+import { slide as Menu } from "react-burger-menu";
+import { Link } from "react-router-dom";
 
-export function Sidebar(){
+export default function Sidebar({ menuItems }) {
+  const [isOpen, setIsOpen] = React.useState(false);
 
-    return(
-            <aside id="side">
-                <nav>
-                    <Link to = "">Rastreio</Link>
-                    <Link to = "/login">Login</Link>
-                    <Link to = "">Cadastro</Link>
-                </nav>
-            </aside>
-    )
+  useEffect(() => {
+    if (isOpen) {
+      document.body.classList.add("sidebar-open");
+    } else {
+      document.body.classList.remove("sidebar-open");
+    }
+    return () => document.body.classList.remove("sidebar-open");
+  }, [isOpen]);
+
+  return (
+    <Menu
+      right
+      isOpen={isOpen}
+      onStateChange={(state) => setIsOpen(state.isOpen)}
+    >
+      <Link className="menu-item" to ="/">
+        Login
+      </Link>
+
+      <Link className="menu-item" to ="/register">
+        Cadastro ADM
+      </Link>
+
+      <Link className="menu-item" to ="/tracking-user">
+        Rastreio
+      </Link>
+    </Menu>
+  );
 }
