--- frontend/src/containers/components/footer/Footer.css
@@ -1,25 +1,0 @@
-#copyright {
-  display: block;
-  font-size: 15px;
-}
-
-footer {
-  display: flex;
-  height: 80px;
-  width: 100vw;
-  justify-items: center;
-  align-items: center;
-  flex-direction: column;
-  border-top: 3px solid rgb(60, 56, 56);
-  background-color: #ffffff;
-}
-
-#info {
-  display: flex;
-  width: 75%;
-  height: auto;
-  justify-items: center;
-  align-items: center;
-  align-content: center;
-  flex-direction: column;
-}
