--- frontend/src/containers/components/footer/Footer.jsx
@@ -1,11 +1,0 @@
-import "./Footer.css";
-
-export function Footer() {
-  return (
-    <footer>
-      <section id="info">
-        <h2 id="copyright">@Copyright - Todos os direitos reservados.</h2>
-      </section>
-    </footer>
-  );
-}
