+++ frontend/src/pages/atualizacaoid/Atualizacaoid.jsx
@@ -1,6 +1,7 @@
 import { Header } from "../../components/header/Header"
 import { Footer } from "../../components/footer/Footer"
 import {Lock} from "lucide-react"
+import SidebarAdm from "../../components/sidebarAdm/SidebarAdm";
 import './Atualizacaoid.css'
 
 export function Atualizacaoid(){
@@ -8,6 +9,7 @@ export function Atualizacaoid(){
     return (
         <>
         <Header />
+        <SidebarAdm/>
             <div className="fundoAtualiza2">
                 <form className="alteracao">
                         <div className="alterainputs">
