+++ frontend/src/pages/atualizacaomed/Atualizacaomed.jsx
@@ -1,5 +1,6 @@
 import { Header } from "../../components/header/Header"
 import { Footer } from "../../components/footer/Footer"
+import SidebarAdm from "../../components/sidebarAdm/SidebarAdm";
 import './Atualizacaomed.css'
 
 export function Atualizacaomed(){
@@ -7,6 +8,7 @@ export function Atualizacaomed(){
     return (
         <>
         <Header />
+        <SidebarAdm/>
             <div className="fundoAtualiza1">
                 <form className="rastreio">
                     <h3>Informe o código de rastreio:</h3>
