+++ frontend/src/pages/registerAdm/Register.jsx
@@ -1,7 +1,7 @@
 import "./Register.css";
-import { Header } from "../components/header/Header";
-import Sidebar from "../components/sidebar/Sidebar";
-import { Footer } from "../components/footer/Footer";
+import { Header } from "../../components/header/Header";
+import Sidebar from "../../components/sidebar/Sidebar";
+import { Footer } from "../../components/footer/Footer";
 
 function TelaCadastro() {
   return (
