+++ frontend/src/pages/resultrackingUser/ResultUser.jsx
@@ -1,16 +1,16 @@
-import "./ResultAdm.css";
-import { Header } from "../components/header/Header";
-import SidebarAdm from "../components/sidebarAdm/SidebarAdm";
-import { Footer } from "../components/footer/Footer";
-import Timeline from "../components/timeline/Timeline";
+import "./ResultUser.css";
+import { Header } from "../../components/header/Header";
+import Sidebar from "../../components/sidebar/Sidebar";
+import { Footer } from "../../components/footer/Footer";
+import Timeline from "../../components/timeline/Timeline";
 
-function ResultA() {
+function Result() {
   return (
-    <div className="container-resultAdm">
-      <div className="contentAdm">
+    <div className="container-resultUser">
+      <div className="contentUser">
         <h1>Rastreio - Pedido: ABC00000000000</h1>
         <Timeline />
-        <div className="deliveryAdm">
+        <div className="delivery">
           <h3>26/12/2024 10:47 Objeto postado</h3>
           <h3>26/12/2024 10:47 Objeto postado</h3>
           <h3>26/12/2024 10:47 Objeto postado</h3>
@@ -21,17 +21,17 @@ function ResultA() {
   );
 }
 
-function ResultAdm() {
+function ResultUser() {
   return (
     <div style={{ display: "flex", flexDirection: "column", height: "100vh" }}>
       <Header />
       <div style={{ display: "flex", flex: 1 }}>
-        <SidebarAdm />
-        <ResultA />
+        <Sidebar />
+        <Result />
       </div>
       <Footer />
     </div>
   );
 }
 
-export default ResultAdm;
+export default ResultUser;
