+++ frontend/src/pages/resultranckingAdm/ResultAdm.jsx
@@ -1,16 +1,16 @@
-import "./ResultUser.css";
-import { Header } from "../components/header/Header";
-import Sidebar from "../components/sidebar/Sidebar";
-import { Footer } from "../components/footer/Footer";
-import Timeline from "../components/timeline/Timeline";
+import "./ResultAdm.css";
+import { Header } from "../../components/header/Header";
+import SidebarAdm from "../../components/sidebarAdm/SidebarAdm";
+import { Footer } from "../../components/footer/Footer";
+import Timeline from "../../components/timeline/Timeline";
 
-function Result() {
+function ResultA() {
   return (
-    <div className="container-resultUser">
-      <div className="contentUser">
+    <div className="container-resultAdm">
+      <div className="contentAdm">
         <h1>Rastreio - Pedido: ABC00000000000</h1>
         <Timeline />
-        <div className="delivery">
+        <div className="deliveryAdm">
           <h3>26/12/2024 10:47 Objeto postado</h3>
           <h3>26/12/2024 10:47 Objeto postado</h3>
           <h3>26/12/2024 10:47 Objeto postado</h3>
@@ -21,17 +21,17 @@ function Result() {
   );
 }
 
-function ResultUser() {
+function ResultAdm() {
   return (
     <div style={{ display: "flex", flexDirection: "column", height: "100vh" }}>
       <Header />
       <div style={{ display: "flex", flex: 1 }}>
-        <Sidebar />
-        <Result />
+        <SidebarAdm />
+        <ResultA />
       </div>
       <Footer />
     </div>
   );
 }
 
-export default ResultUser;
+export default ResultAdm;
