+++ frontend/src/pages/trackingUser/TrackingUser.jsx
@@ -1,7 +1,7 @@
 import "./TranckingUser.css";
-import { Header } from "../components/header/Header";
-import Sidebar from "../components/sidebar/Sidebar";
-import { Footer } from "../components/footer/Footer";
+import { Header } from "../../components/header/Header";
+import Sidebar from "../../components/sidebar/Sidebar";
+import { Footer } from "../../components/footer/Footer";
 
 function TelaBusca() {
   return (
